+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/universal/mapper/CodedRegionReferenceMapper.java
@@ -25,19 +25,19 @@ public class CodedRegionReferenceMapper {
             TRegion region = RectangleRegionMapper.toRectangleRegionDto(simpleRegionByRectangle.getRegion());
             codedRegionReference.setRegion(region);
             codedRegionReference.setRegionId(simpleRegionByRectangle.getRegion().getRegionId());
-            // set padding here
+            codedRegionReference.setPadding(simpleRegionByRectangle.getPadding());
         } else if (getSimpleRegion instanceof SimpleRegionByElement) {
             SimpleRegionByElement simpleRegionByElement = (SimpleRegionByElement) getSimpleRegion;
             TRegion region = ElementRegionMapper.toElementRegionDto(simpleRegionByElement.getElement());
             codedRegionReference.setRegion(region);
             codedRegionReference.setRegionId(simpleRegionByElement.getRegionId());
-            // set padding here
+            codedRegionReference.setPadding(simpleRegionByElement.getPadding());
         } else if (getSimpleRegion instanceof SimpleRegionBySelector) {
             SimpleRegionBySelector simpleRegionBySelector = (SimpleRegionBySelector) getSimpleRegion;
             TRegion region = SelectorRegionMapper.toSelectorRegionDto(simpleRegionBySelector.getSelector());
             codedRegionReference.setRegion(region);
             codedRegionReference.setRegionId(simpleRegionBySelector.getRegionId());
-            // set padding here
+            codedRegionReference.setPadding(simpleRegionBySelector.getPadding());
         }
 
         return codedRegionReference;